OnboxAPI/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
filelock==3.0.12
httpie==2.2.0
idna==2.10
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
